refactor: rename PasswordPopup to UnlockStrongholdPopup
