+++ packages/shared/components/popups/UnlockStrongholdPopup.svelte
